requirements.txt
customtkinter
PyYAML
requests
vcolorpicker
psutil
GPUtil
wmi